requirements.txt
# Core dependencies
requests==2.31.0
feedparser==6.0.10
Jinja2==3.1.2
python-dotenv==1.0.0

# Google Calendar API (optional)
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0

# Development dependencies
pytest==7.4.3
black==23.12.1
flake8==7.0.0